// src/types/api/guild.ts
export declare class GuildConfiguration {
  guildId: string;
  messagesImage: string;
  messagesFooterName: string;
  scanMainRegionName: string;
  mainChannelName: string;
  adminChannelName: string;
  emergencyChannelName: string;
  acknowledged: string;
  activityNoticeMinPlayers: number;
  activityNoticeMessages: Array<string>;
  commandsPrefix: string;
}